Adress PR reviews
